requirements.txt
# Core Flask dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2

# Database
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7  # PostgreSQL adapter for production

# Email
Flask-Mail==0.9.1
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1

# PDF generation
pdfkit==1.0.0
stripe==12.2.0  # Updated to latest version

# Environment and configuration
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1

# OAuth and authentication
Flask-Dance[google]==7.0.0
oauthlib==3.2.2

# Production server
gunicorn==21.2.0

# Testing
pytest==7.4.2
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development and code quality
black==23.9.1
flake8==6.1.0
mypy==1.5.1

# Security
cryptography==41.0.4

